refactor: OrderController 삭제 및 LottoOrder 클래스명 변경
