+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.controller.OrderController;
+import lotto.domain.LottoGame;
 import lotto.dto.LottoResultDto;
 import lotto.dto.LottoTicketsDto;
 import lotto.view.InputView;
@@ -10,10 +10,9 @@ import lotto.view.OutputView;
 public class Application {
 
     public static void main(String[] args) {
-        OrderController orderController = new OrderController();
-        orderController.takeOrder(InputView.requestMoney(), InputView.requestManualLotto());
+        LottoGame lottoGame = new LottoGame(InputView.requestMoney(), InputView.requestManualLotto());
 
-        LottoController lottoController = new LottoController(orderController.getLottoOrder());
+        LottoController lottoController = new LottoController(lottoGame);
         LottoTicketsDto lottoTicketsDto = lottoController.publishLottoTickets();
         OutputView.displayLottoTickets(lottoTicketsDto);
