+++ src/main/java/lotto/domain/LottoGame.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import java.util.ArrayList;
 import java.util.List;
 
-public class LottoOrder {
+public class LottoGame {
 
     public static final int LOTTO_PRICE = 1000;
     private static final String ERROR_ORDER_NUMBER = "금액이 부족하여 주문한 수만큼 로또를 구매할 수 없습니다.";
@@ -11,7 +11,7 @@ public class LottoOrder {
     private Money money;
     private final List<List<Integer>> manualNumbers;
 
-    public LottoOrder(int inputMoney, List<List<Integer>> numbers) {
+    public LottoGame(int inputMoney, List<List<Integer>> numbers) {
         this.money = new Money(inputMoney, LOTTO_PRICE);
         this.manualNumbers = numbers;
         validateOrderNumber();
