+++ src/test/java/lottoTest/domain/LottoGameTest.java
@@ -5,12 +5,12 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.domain.LottoOrder;
+import lotto.domain.LottoGame;
 import org.junit.jupiter.api.Test;
 
 
 @SuppressWarnings("NonAsciiCharacters")
-class LottoOrderTest {
+class LottoGameTest {
 
     @Test
     void 수동_로또_주문수에_비해서_금액이_부족한_경우_테스트() {
@@ -18,7 +18,7 @@ class LottoOrderTest {
         manualNumbers.add(List.of(1, 2, 3, 4, 5, 6));
         manualNumbers.add(List.of(7, 8, 9, 10, 11, 12));
 
-        assertThatThrownBy(() -> new LottoOrder(1000, manualNumbers))
+        assertThatThrownBy(() -> new LottoGame(1000, manualNumbers))
                 .isInstanceOf(RuntimeException.class)
                 .hasMessageContaining("금액이 부족하여 주문한 수만큼 로또를 구매할 수 없습니다.");
     }
@@ -27,9 +27,9 @@ class LottoOrderTest {
     void 수동_로또를_주문하고_남은금액으로_자동_로또_개수를_게산하는_기능_테스트() {
         List<List<Integer>> manualNumbers = new ArrayList<>();
         manualNumbers.add(List.of(1, 2, 3, 4, 5, 6));
-        LottoOrder lottoOrder = new LottoOrder(2000, manualNumbers);
-        lottoOrder.billingManualLottoOrder();
+        LottoGame lottoGame = new LottoGame(2000, manualNumbers);
+        lottoGame.billingManualLottoOrder();
 
-        assertThat(lottoOrder.getNumberOfAutoLotto()).isEqualTo(1);
+        assertThat(lottoGame.getNumberOfAutoLotto()).isEqualTo(1);
     }
 }
