+++ src/test/java/lottoTest/domain/MoneyTest.java
@@ -1,6 +1,6 @@
 package lottoTest.domain;
 
-import static lotto.domain.LottoOrder.LOTTO_PRICE;
+import static lotto.domain.LottoGame.LOTTO_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
